OPA v1.18.2
Also:
- Remove codecov action as it's broken and we don't really use that
- Add simple benchmarks for each of our custom builtin functions
- Disable exhaustive linter as it's annoying
- Some more interning
