--- .github/codecov.yml
@@ -1,1 +1,0 @@
-comment: false
